Add live currency exchanger

+++ app/src/main/java/com/oleksandrstefanyshyn/demoapp/CurrenciesApi.kt
@@ -3,7 +3,6 @@ package com.oleksandrstefanyshyn.demoapp
 import io.reactivex.Single
 import retrofit2.http.GET
 import retrofit2.http.Query
-import java.math.BigDecimal
 
 interface CurrenciesApi {
 
@@ -14,4 +13,4 @@ interface CurrenciesApi {
     fun exangeRates(@Query("base") base: String): Single<Currencies>
 }
 
-data class Currencies(val rates: Map<String, BigDecimal>)
+data class Currencies(val rates: Map<String, Double>)
